+++ lib/service/Appointments.dart
@@ -19,7 +19,8 @@ import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/model/wellness/Appointment.dart';
+import 'package:illinois/model/Appointment.dart';
+import 'package:illinois/service/Gateway.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
@@ -482,6 +483,9 @@ class Appointments with Service implements NotificationsListener {
 
   // Debug
 
+  static String get externalAuthorizationHeaderKey => Gateway.ExternalAuthorizationHeaderKey;
+  Map<String, String?> get externalAuthorizationHeader => Gateway().externalAuthorizationHeader;
+
   bool? get _useSampleData => Storage().debugUseSampleAppointments;
 
   // Providers
@@ -489,7 +493,7 @@ class Appointments with Service implements NotificationsListener {
   Future<List<AppointmentProvider>?> loadProviders() async {
     if (_useSampleData != true) {
       String? url = "${Config().appointmentsUrl}/services/providers";
-      http.Response? response = await Network().get(url, auth: Auth2());
+      http.Response? response = await Network().get(url, headers: externalAuthorizationHeader, auth: Auth2());
       return (response?.statusCode == 200) ? AppointmentProvider.listFromJson(JsonUtils.decodeList(response?.body)) : null;
     }
     else {
@@ -499,10 +503,10 @@ class Appointments with Service implements NotificationsListener {
   }
 
   List<AppointmentProvider> get _sampleProviders =>  <AppointmentProvider>[
-    AppointmentProvider(id: '1', name: 'McKinley'),
-    AppointmentProvider(id: '2', name: 'Grainger'),
-    AppointmentProvider(id: '3', name: 'Lorem Ipsum'),
-    AppointmentProvider(id: '4', name: 'Sit Dolor Amet'),
+    AppointmentProvider(id: '1', name: 'McKinley', supportsSchedule: true, supportsReschedule: true, supportsCancel: true),
+    AppointmentProvider(id: '2', name: 'Grainger', supportsSchedule: false, supportsReschedule: false, supportsCancel: false),
+    AppointmentProvider(id: '3', name: 'Lorem Ipsum', supportsSchedule: true, supportsReschedule: true, supportsCancel: false),
+    AppointmentProvider(id: '4', name: 'Sit Dolor Amet', supportsSchedule: true, supportsReschedule: false, supportsCancel: true),
   ];
 
   // Units
@@ -510,7 +514,7 @@ class Appointments with Service implements NotificationsListener {
   Future<List<AppointmentUnit>?> loadUnits({ required String providerId }) async {
     if (_useSampleData != true) {
       String? url = "${Config().appointmentsUrl}/services/units?providers-ids=$providerId";
-      http.Response? response = await Network().get(url, auth: Auth2());
+      http.Response? response = await Network().get(url, headers: externalAuthorizationHeader, auth: Auth2());
       return (response?.statusCode == 200) ? AppointmentUnit.listFromJson(JsonUtils.decodeList(response?.body)) : null;
     }
     else {
@@ -531,13 +535,14 @@ class Appointments with Service implements NotificationsListener {
   Future<List<AppointmentPerson>?> loadPersons({ required String providerId, required String unitId }) async {
     if (_useSampleData != true) {
       String? url = "${Config().appointmentsUrl}/services/people";
-      Map<String, String> headers = {
+      Map<String, String?> headers = {
         'Content-Type': 'application/json'
       };
       String? post = JsonUtils.encode([{
         'provider_id': providerId,
         'unit_ids': [ unitId ],
       }]);
+      headers.addAll(externalAuthorizationHeader);
       http.Response? response = await Network().get(url, body: post, headers: headers, auth: Auth2());
       return (response?.statusCode == 200) ? AppointmentPerson.listFromJson(JsonUtils.decodeList(response?.body)) : null;
     }
@@ -564,8 +569,18 @@ class Appointments with Service implements NotificationsListener {
     if (_useSampleData != true) {
       int startTime = DateUtils.dateOnly(dateLocal).millisecondsSinceEpoch;
       int endTime = startTime + 86400000; // 1 day in milliseconds = 24 * 60 * 60 * 1000
-      String? url = "${Config().appointmentsUrl}/services/slots?provider-id=$providerId&unit-id=$unitId&person-id=$personId&start-time=$startTime&end-time=$endTime";
-      http.Response? response = await Network().get(url, auth: Auth2());
+      String urlParams = 'start-time=$startTime&end-time=$endTime';
+      if (providerId != null) {
+        urlParams += "&provider-id=$providerId";
+      }
+      if (unitId != null) {
+        urlParams += "&unit-id=$unitId";
+      }
+      if (personId != null) {
+        urlParams += "&person-id=$personId";
+      }
+      String? url = "${Config().appointmentsUrl}/services/slots?$urlParams";
+      http.Response? response = await Network().get(url, headers: externalAuthorizationHeader, auth: Auth2());
       return (response?.statusCode == 200) ? AppointmentTimeSlotsAndQuestions.fromJson(JsonUtils.decodeMap(response?.body)) : null;
     }
     else {
@@ -615,8 +630,11 @@ class Appointments with Service implements NotificationsListener {
 
   Future<List<Appointment>?> loadAppointments({String? providerId}) async {
     if (_useSampleData != true) {
-      String? url = "${Config().appointmentsUrl}/services/appointments?providers-ids=$providerId";
-      http.Response? response = await Network().get(url, auth: Auth2());
+      String url = "${Config().appointmentsUrl}/services/v2/appointments";
+      if (providerId != null) {
+        url += "?providers-ids=$providerId";
+      }
+      http.Response? response = await Network().get(url, headers: externalAuthorizationHeader, auth: Auth2());
       return (response?.statusCode == 200) ? Appointment.listFromJson(JsonUtils.decodeList(response?.body)) : null;
     }
     else {
@@ -626,7 +644,11 @@ class Appointments with Service implements NotificationsListener {
         return _sampleAppointments(provider: provider);
       }
       else {
-        return <Appointment>[];
+        List<Appointment> result = <Appointment>[];
+        for(AppointmentProvider provider in _sampleProviders) {
+          result.addAll(_sampleAppointments(provider: provider));
+        }
+        return result;
       }
     }
   }
@@ -662,7 +684,7 @@ class Appointments with Service implements NotificationsListener {
     List<AppointmentPerson> persons = _samplePersons;
     AppointmentPerson person = persons[Random().nextInt(persons.length)];
     
-    bool cancelled = ((Random().nextInt(3) % 5) == 0);
+    bool cancelled = (provider.supportsCancel == true) && ((Random().nextInt(3) % 5) == 0);
 
     DateTime startTimeUtc = DateTime(day.year, day.month, day.day, Random().nextInt(8) + 8, 30).toUtc();
     DateTime endTimeUtc = startTimeUtc.add(Duration(minutes: 30));
@@ -711,9 +733,10 @@ class Appointments with Service implements NotificationsListener {
   }) async {
     if (_useSampleData != true) {
       String? url = "${Config().appointmentsUrl}/services/appointments";
-      Map<String, String> headers = {
+      Map<String, String?> headers = {
         'Content-Type': 'application/json'
       };
+      headers.addAll(externalAuthorizationHeader);
       String? post = JsonUtils.encode({
         'provider_id': provider?.id,
         'unit_id': unit?.id,
@@ -747,9 +770,10 @@ class Appointments with Service implements NotificationsListener {
   }) async {
     if (_useSampleData != true) {
       String? url = "${Config().appointmentsUrl}/services/appointments/${appointment.id}";
-      Map<String, String> headers = {
+      Map<String, String?> headers = {
         'Content-Type': 'application/json'
       };
+      headers.addAll(externalAuthorizationHeader);
       String? post = JsonUtils.encode({
         'type': appointmentTypeToString(type),
         'time': timeSlot?.startTimeUtc?.millisecondsSinceEpoch,
@@ -776,7 +800,7 @@ class Appointments with Service implements NotificationsListener {
   Future<Appointment?> cancelAppointment(Appointment appointment) async {
     if (_useSampleData != true) {
         String? url = "${Config().appointmentsUrl}/services/appointments/${appointment.id}";
-        http.Response? response = await Network().delete(url, auth: Auth2());
+        http.Response? response = await Network().delete(url, headers: externalAuthorizationHeader, auth: Auth2());
         if (response?.statusCode == 200) {
           return Appointment.fromOther(appointment, cancelled: true);
         }